Move language and linter under wrapper pkg
